docs(tour): second-round review polish on Hamiltonian path
Follow-up to the API-change commit, addressing reviewer polish notes.

- Replace stale "null is returned" language in BacktrackingHamiltonianPath
  with the tri-state PROVEN_ABSENT wording now that getPath returns a
  HamiltonianPathSearchResult.

- Soften the literature paragraph so it no longer implies the code is a
  verbatim implementation of Rubin's algorithm: it is a straightforward
  exact DFS/backtracking solver using standard pruning and
  candidate-ordering ideas, with the references given as background.

- Correct the DagHamiltonianPath cyclic-input Javadoc: it now states that
  cyclic input raises an IllegalArgumentException *wrapping* the
  NotDirectedAcyclicGraphException thrown by TopologicalOrderIterator,
  matching the catch-and-rethrow in the code.

- Make the result implementation genuinely non-public: move the former
  nested HamiltonianPathSearchResult.Default (implicitly public as an
  interface member) to a package-private top-level class
  DefaultHamiltonianPathSearchResult in the same package. The interface
  factories construct it; it is no longer part of the public API surface.

No behavioural change. jgrapht-core FastTestSuite (5255 tests) and the
checkstyle profile both pass.
